Update getting of counts of returned records

+++ library/core/src/main/java/cz/matfyz/core/adminer/DataResponse.java
@@ -8,7 +8,7 @@ import java.util.List;
 public abstract class DataResponse {
     private Metadata metadata;
 
-    protected DataResponse(int itemCount, List<String> propertyNames) {
+    protected DataResponse(long itemCount, List<String> propertyNames) {
         this.metadata = new Metadata(itemCount, propertyNames);
     }
 
@@ -19,7 +19,7 @@ public abstract class DataResponse {
     /**
      * Represents metadata for the response.
      */
-    public record Metadata (int itemCount, List<String> propertyNames) {}
+    public record Metadata (long itemCount, List<String> propertyNames) {}
 
     public abstract String getType();
